+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2026-02-08
+
+### Changed
+- Default authentication keyword from `Bearer` to `Api-Key` to align with upstream `djangorestframework-api-key`
+- Keyword matching is now case-insensitive
+
+[0.2.0]: https://github.com/frankapps-io/drf-scoped-permissions/releases/tag/v0.2.0
+
 ## [0.1.1] - 2025-02-07
 
 ### Added